[Update] Location 기능 추상화
Location 관련 기능을 좀 더 쉽게 사용하기 위해 추상화함
