+++ presentation/src/main/java/com/mohaeyo/mohae/base/EndPointLocationFragment.kt
@@ -1,0 +1,40 @@
+package com.mohaeyo.mohae.base
+
+import android.content.Context
+import androidx.activity.OnBackPressedCallback
+import androidx.databinding.ViewDataBinding
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import io.reactivex.subjects.BehaviorSubject
+import io.reactivex.subjects.Subject
+import org.jetbrains.anko.support.v4.toast
+
+abstract class EndPointLocationFragment<V: ViewDataBinding>: BaseLocationFragment<V>() {
+    val backButtonSubject: Subject<Long> =
+        BehaviorSubject.createDefault(0L)
+            .toSerialized()
+
+    private val backButtonSubjectDisposable: Disposable = backButtonSubject
+        .buffer(2, 1)
+        .observeOn(AndroidSchedulers.mainThread())
+        .subscribe {
+            if (it[1] - it[0] <= 1500) activity?.finish()
+            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
+        }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+
+        requireActivity().onBackPressedDispatcher.addCallback(
+            this, object: OnBackPressedCallback(true) {
+                override fun handleOnBackPressed() {
+                    backButtonSubject.onNext(System.currentTimeMillis())
+                }
+            })
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        backButtonSubjectDisposable.dispose()
+    }
+}
